fixing the code functionality so it displays code Snippet and imports all elements in App.jsx, fixed main.jsx too, added github folder workflows as a start for deployment
