+++ src/main.jsx
@@ -1,10 +1,23 @@
 import React from 'react';
-import { createRoot } from 'react-dom/client';
-import App from './App';
-import ErrorBoundary from './components/ErrorBoundary/ErrorBoundary';
+import ReactDOM from 'react-dom/client';
 
-createRoot(document.getElementById('root')).render(
-  <ErrorBoundary>
+import App from './App';
+
+
+
+const root = ReactDOM.createRoot(document.getElementById('root'));
+
+root.render(
+  <React.StrictMode>
     <App />
-  </ErrorBoundary>
+  </React.StrictMode>
 );
+
+
+// import { createRoot } from 'react-dom/client';
+// import ErrorBoundary from './ErrorBoundary/ErrorBoundary';
+// createRoot(document.getElementById('root')).render(
+//   <ErrorBoundary>
+//     <App />
+//   </ErrorBoundary>
+// );
